requirements.txt
Flask==2.1.0
numpy==1.16.6
python-dotenv==0.15.0
requests==2.25.0
urllib3==1.26.6
